Label reset + Saved cell labels
Added a reset for all labels when another new entry is being added. Previously when tapping the addEntryButton the information entered in the last . added entry would still occupy the labels.

Location label information was not properly displayed yet in the saved entry cells as they were stored but never loaded from CoreData.